一天学习一次 多线程 netty 入门 用Pipline改造

+++ src/main/java/Z_0019_Netty入门IO通讯/S9_Netty客户端与服务端收发消息/F1_序列化算法类型.java
@@ -9,10 +9,10 @@ public interface F1_序列化算法类型 {
     /**
      * JSON 算法类型
      */
-    int JSON = 1;
+    byte JSON = 1;
 
     /**
      * JDK 序列化算法类型
      */
-    int JDK = 2;
+    byte JDK = 2;
 }
