+++ src/main/java/Z_0019_Netty入门IO通讯/S9_Netty客户端与服务端收发消息/F5_编码.java
@@ -11,7 +11,7 @@ public class F5_编码 {
 
     public static ByteBuf 编码(ByteBufAllocator 载体, F1_抽象数据包 抽象数据包) {
         F3_JSON序列化实现 json序列化实现 = F2_序列化接口.默认实现;
-        int 获取算法类型 = json序列化实现.获取算法类型();
+        byte 获取算法类型 = json序列化实现.获取算法类型();
 
         byte[] 序列化好的数据 = json序列化实现.序列化(抽象数据包);
         int 数据长度 = 序列化好的数据.length;
@@ -19,7 +19,7 @@ public class F5_编码 {
         ByteBuf 数据载体 = 载体.ioBuffer();
         数据载体.writeInt(F1_抽象数据包.魔数);
         数据载体.writeByte(抽象数据包.版本);
-        数据载体.writeInt(获取算法类型);
+        数据载体.writeByte(获取算法类型);
         数据载体.writeByte(抽象数据包.指令());
         数据载体.writeInt(数据长度);
         数据载体.writeBytes(序列化好的数据);
